Rename measure packages
