+++ examples/src/main/java/com/opengamma/strata/examples/finance/CurveScenarioExample.java
@@ -44,7 +44,7 @@ import com.opengamma.strata.market.curve.Curve;
 import com.opengamma.strata.market.curve.CurveParallelShifts;
 import com.opengamma.strata.measure.Measures;
 import com.opengamma.strata.measure.StandardComponents;
-import com.opengamma.strata.measure.calc.curve.AnyCurveFilter;
+import com.opengamma.strata.measure.curve.AnyCurveFilter;
 import com.opengamma.strata.product.Trade;
 import com.opengamma.strata.product.TradeAttributeType;
 import com.opengamma.strata.product.TradeInfo;
