+++ modules/measure/src/main/java/com/opengamma/strata/measure/StandardComponents.java
@@ -14,15 +14,15 @@ import com.opengamma.strata.calc.marketdata.MarketDataFunction;
 import com.opengamma.strata.calc.marketdata.ObservableDataProvider;
 import com.opengamma.strata.calc.marketdata.TimeSeriesProvider;
 import com.opengamma.strata.calc.runner.CalculationFunctions;
-import com.opengamma.strata.measure.calc.curve.CurveGroupMarketDataFunction;
-import com.opengamma.strata.measure.calc.curve.CurveInputsMarketDataFunction;
-import com.opengamma.strata.measure.calc.curve.CurveMarketDataFunction;
-import com.opengamma.strata.measure.calc.fx.FxRateMarketDataFunction;
 import com.opengamma.strata.measure.credit.CdsCalculationFunction;
+import com.opengamma.strata.measure.curve.CurveGroupMarketDataFunction;
+import com.opengamma.strata.measure.curve.CurveInputsMarketDataFunction;
+import com.opengamma.strata.measure.curve.CurveMarketDataFunction;
 import com.opengamma.strata.measure.deposit.TermDepositCalculationFunction;
 import com.opengamma.strata.measure.dsf.DsfCalculationFunction;
 import com.opengamma.strata.measure.fra.FraCalculationFunction;
 import com.opengamma.strata.measure.fx.FxNdfCalculationFunction;
+import com.opengamma.strata.measure.fx.FxRateMarketDataFunction;
 import com.opengamma.strata.measure.fx.FxSingleCalculationFunction;
 import com.opengamma.strata.measure.fx.FxSwapCalculationFunction;
 import com.opengamma.strata.measure.index.IborFutureCalculationFunction;
