--- modules/measure/src/main/java/com/opengamma/strata/measure/calc/curve/package-info.java
@@ -1,11 +1,0 @@
-/**
- * Copyright (C) 2015 - present by OpenGamma Inc. and the OpenGamma group of companies
- *
- * Please see distribution for license.
- */
-
-/**
- * Integration code that allows strata-calc to use and calibrate curves.
- */
-package com.opengamma.strata.measure.calc.curve;
-
