+++ modules/measure/src/main/java/com/opengamma/strata/measure/curve/CurveNameFilter.java
@@ -3,7 +3,7 @@
  *
  * Please see distribution for license.
  */
-package com.opengamma.strata.measure.calc.curve;
+package com.opengamma.strata.measure.curve;
 
 import java.util.Map;
 import java.util.NoSuchElementException;
