+++ modules/measure/src/main/java/com/opengamma/strata/measure/curve/RootFinderConfig.java
@@ -3,7 +3,7 @@
  *
  * Please see distribution for license.
  */
-package com.opengamma.strata.measure.calc.curve;
+package com.opengamma.strata.measure.curve;
 
 import java.io.Serializable;
 import java.util.Map;
