+++ modules/measure/src/test/java/com/opengamma/strata/measure/curve/CurveNameFilterTest.java
@@ -3,8 +3,9 @@
  *
  * Please see distribution for license.
  */
-package com.opengamma.strata.measure.calc.curve;
+package com.opengamma.strata.measure.curve;
 
+import static com.opengamma.strata.collect.TestHelper.coverBeanEquals;
 import static com.opengamma.strata.collect.TestHelper.coverImmutableBean;
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -14,25 +15,35 @@ import com.opengamma.strata.data.scenario.MarketDataBox;
 import com.opengamma.strata.market.curve.ConstantCurve;
 import com.opengamma.strata.market.curve.Curve;
 import com.opengamma.strata.market.curve.CurveId;
+import com.opengamma.strata.market.curve.CurveName;
 
 /**
- * Test {@link AnyCurveFilter}.
+ * Test {@link CurveNameFilter}.
  */
 @Test
-public class AnyCurveFilterTest {
+public class CurveNameFilterTest {
 
   public void match() {
-    AnyCurveFilter test = AnyCurveFilter.INSTANCE;
+    CurveNameFilter test = CurveNameFilter.of(CurveName.of("name"));
     assertThat(test.getMarketDataIdType()).isEqualTo(CurveId.class);
     CurveId id = CurveId.of("group", "name");
     Curve curve = ConstantCurve.of("name", 1);
     assertThat(test.matches(id, MarketDataBox.ofSingleValue(curve))).isTrue();
   }
 
+  public void noMatch() {
+    CurveNameFilter test = CurveNameFilter.of(CurveName.of("name"));
+    CurveId id = CurveId.of("group", "name");
+    Curve curve = ConstantCurve.of("notCurveName", 1);
+    assertThat(test.matches(id, MarketDataBox.ofSingleValue(curve))).isFalse();
+  }
+
   //-------------------------------------------------------------------------
   public void coverage() {
-    AnyCurveFilter test = AnyCurveFilter.INSTANCE;
+    CurveNameFilter test = CurveNameFilter.of(CurveName.of("curveName1"));
     coverImmutableBean(test);
+    CurveNameFilter test2 = CurveNameFilter.of(CurveName.of("curveName2"));
+    coverBeanEquals(test, test2);
   }
 
 }
