+++ modules/measure/src/test/java/com/opengamma/strata/measure/curve/TestMarketDataMap.java
@@ -3,7 +3,7 @@
  *
  * Please see distribution for license.
  */
-package com.opengamma.strata.measure.calc.curve;
+package com.opengamma.strata.measure.curve;
 
 import static com.opengamma.strata.collect.Guavate.toImmutableSet;
