+++ modules/measure/src/test/java/com/opengamma/strata/measure/fra/FraCalculationFunctionTest.java
@@ -41,7 +41,7 @@ import com.opengamma.strata.market.observable.IndexQuoteId;
 import com.opengamma.strata.market.param.CurrencyParameterSensitivities;
 import com.opengamma.strata.market.sensitivity.PointSensitivities;
 import com.opengamma.strata.measure.Measures;
-import com.opengamma.strata.measure.calc.curve.TestMarketDataMap;
+import com.opengamma.strata.measure.curve.TestMarketDataMap;
 import com.opengamma.strata.measure.rate.RatesMarketDataLookup;
 import com.opengamma.strata.pricer.fra.DiscountingFraProductPricer;
 import com.opengamma.strata.pricer.fra.FraDummyData;
