+++ modules/measure/src/test/java/com/opengamma/strata/measure/fx/FxNdfCalculationFunctionTest.java
@@ -41,7 +41,7 @@ import com.opengamma.strata.market.curve.Curves;
 import com.opengamma.strata.market.param.CurrencyParameterSensitivities;
 import com.opengamma.strata.market.sensitivity.PointSensitivities;
 import com.opengamma.strata.measure.Measures;
-import com.opengamma.strata.measure.calc.curve.TestMarketDataMap;
+import com.opengamma.strata.measure.curve.TestMarketDataMap;
 import com.opengamma.strata.measure.rate.RatesMarketDataLookup;
 import com.opengamma.strata.pricer.fx.DiscountingFxNdfProductPricer;
 import com.opengamma.strata.pricer.rate.RatesProvider;
