+++ modules/measure/src/test/java/com/opengamma/strata/measure/fx/FxRateConfigTest.java
@@ -3,7 +3,7 @@
  *
  * Please see distribution for license.
  */
-package com.opengamma.strata.measure.calc.fx;
+package com.opengamma.strata.measure.fx;
 
 import static com.opengamma.strata.collect.TestHelper.assertThrowsIllegalArg;
 import static org.assertj.core.api.Assertions.assertThat;
