+++ modules/measure/src/test/java/com/opengamma/strata/measure/index/IborFutureCalculationFunctionTest.java
@@ -38,7 +38,7 @@ import com.opengamma.strata.market.curve.Curves;
 import com.opengamma.strata.market.observable.IndexQuoteId;
 import com.opengamma.strata.market.observable.QuoteId;
 import com.opengamma.strata.measure.Measures;
-import com.opengamma.strata.measure.calc.curve.TestMarketDataMap;
+import com.opengamma.strata.measure.curve.TestMarketDataMap;
 import com.opengamma.strata.measure.rate.RatesMarketDataLookup;
 import com.opengamma.strata.pricer.index.DiscountingIborFutureTradePricer;
 import com.opengamma.strata.pricer.rate.RatesProvider;
