+++ modules/measure/src/test/java/com/opengamma/strata/measure/rate/RatesMarketDataLookupTest.java
@@ -49,7 +49,7 @@ import com.opengamma.strata.market.curve.CurveId;
 import com.opengamma.strata.market.curve.CurveName;
 import com.opengamma.strata.market.curve.Curves;
 import com.opengamma.strata.market.observable.IndexQuoteId;
-import com.opengamma.strata.measure.calc.curve.TestMarketDataMap;
+import com.opengamma.strata.measure.curve.TestMarketDataMap;
 import com.opengamma.strata.pricer.SimpleDiscountFactors;
 import com.opengamma.strata.pricer.rate.DiscountIborIndexRates;
 import com.opengamma.strata.pricer.rate.DiscountOvernightIndexRates;
